% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRR-package.R
\docType{package}
\name{DRR-package}
\alias{DRR}
\alias{DRR-package}
\title{Dimensionality Reduction via Regression.}
\description{
DRR implements the Dimensionality Reduction via Regression using
Kernel Ridge Regression. It also adds a faster implementation of
Kernel Ridge regression that can be used with the CVST package.
}
\details{
Funding provided by the Department for Biogeochemical Integration,
Empirical Inference of the Earth System Group, at the Max Plack
Institute for Biogeochemistry, Jena.
}
\references{
Laparra, V., Malo, J., Camps-Valls, G., 2015. Dimensionality
    Reduction via Regression in Hyperspectral Imagery. IEEE Journal
    of Selected Topics in Signal Processing 9,
    1026-1036. doi:10.1109/JSTSP.2015.2417833
Zhang, Y., Duchi, J.C., Wainwright, M.J., 2013. Divide and Conquer
    Kernel Ridge Regression: A Distributed Algorithm with Minimax
    Optimal Rates. arXiv:1305.5029 [cs, math, stat].
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.guido-kraemer.com/software/drr/}
  \item Report bugs at \url{https://github.com/gdkrmr/DRR/issues}
}

}
\author{
\strong{Maintainer}: Guido Kraemer \email{guido.kraemer@uni-leipzig.de} (\href{https://orcid.org/0000-0003-4865-5041}{ORCID})

Authors:
\itemize{
  \item Guido Kraemer \email{guido.kraemer@uni-leipzig.de} (\href{https://orcid.org/0000-0003-4865-5041}{ORCID})
}

}
