% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastKRR.R
\name{makeBlocks}
\alias{makeBlocks}
\title{Fast implementation of Kernel Ridge Regression}
\usage{
makeBlocks(nobs, nblocks)
}
\arguments{
\item{nobs}{total number of observations}

\item{nblocks}{number of blocks}

\item{data}{output of CVST::constructData}

\item{kernel}{kernel function or string, see
  \code{\link[kernlab]{kernel-klass}}
dividing nobs into nblocks parts that have approximately the same size}
}
\value{
vector of integers of length \code{nblocks} that sums up to
  \code{nobs}
}
\description{
To be used as a dropin replacement of .krr to be used with constructFastKRRLearner
}
\details{
implements a divide and conquer scheme that is easier on large datasets
}

